Running Prettier before committing.
