+++ .gitignore
@@ -1,2 +1,3 @@
+**/*.*.swp
 **/.DS_Store
 **/node_modules/
