+++ test-scripts/axe-e2e/index.js
@@ -1,10 +1,12 @@
-var reporter = require('./scripts/e2e-a11y');
+var reporter = require('./scripts/e2e-axe');
 
 if (process.env['NODE_ENV'] === 'report') {
   // Output CSV report in ./csv-reports
   reporter.createCSVReport('./fixtures/sitemap.example.xml');
+} else if (process.env['NODE_ENV'] === 'consoleLight') {
+  // Output to console with light display
+  reporter.scanToConsole('./fixtures/sitemap.example.xml', 'light');
 } else {
-  // Output to console
-  // TODO: Add a light option for console output
-  reporter.scanToConsole('./fixtures/sitemap.example.xml');
+  // Output to console with dark display
+  reporter.scanToConsole('./fixtures/sitemap.example.xml', 'dark');
 }
