+++ test-scripts/axe-e2e/package.json
@@ -15,6 +15,7 @@
   },
   "scripts": {
     "test:console-dark": "node index",
+    "test:console-light": "NODE_ENV=consoleLight node index",
     "test:report": "NODE_ENV=report node index > ./csv-reports/base.csv"
   }
 }
